Agentmanager without cat instance; refactor ws notifications (broken)
