+++ core/cat/memory/working_memory.py
@@ -1,4 +1,3 @@
-import asyncio
 
 class WorkingMemory(dict):
     """Cat's volatile memory.
@@ -19,7 +18,6 @@ class WorkingMemory(dict):
     def __init__(self):
         # The constructor instantiates a `dict` with a 'history' key to store conversation history
         # and the asyncio queue to manage the session notifications
-        self.ws_messages = asyncio.Queue()
         super().__init__(history=[])
 
     def get_user_id(self):
